Move rich renderers to different file
